+++ src-docs/src/views/header/header_example.js
@@ -117,7 +117,7 @@ export const HeaderExample = {
       demo: <HeaderLinks />,
     },
     {
-      title: 'Alerts',
+      title: 'Display header alerts',
       source: [
         {
           type: GuideSectionTypes.JS,