+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Improved `EuiSwitch` a11y by aligning to aria roles ([#2491](https://github.com/elastic/eui/pull/2491))
 - Converted `EuiSwitch` to TypeScript ([#2491](https://github.com/elastic/eui/pull/2491))
 - Added an accessible label-less `EuiSwitch` variation ([#2491](https://github.com/elastic/eui/pull/2491))
-- Updated `EuiHeaderAlert` to add `badge` prop ([#2506](https://github.com/elastic/eui/pull/2506))
+- Added `badge` prop and new styles `EuiHeaderAlert` ([#2506](https://github.com/elastic/eui/pull/2506))
 
 **Bug fixes**